requirements.txt
# Python 3.6.1
Django>=1.11.5
django-cors-headers>=2.1.0
djangorestframework>=3.6.4
Pillow>=4.3.0
geoip2>=2.6.0
